Working through lesson on validation. Updated authorization function

+++ trello_clone/blueprints/cards_bp.py
@@ -2,11 +2,13 @@ from flask import Blueprint, request
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from setup import db
 from models.card import CardSchema, Card
-from auth import admin_required
+from auth import authorize
 from blueprints.comments_bp import comments_bp
 
 
 cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
+
+cards_bp.register_blueprint(comments_bp)
 
 
 # Get all cards
@@ -14,26 +16,27 @@ cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
 @jwt_required()
 def all_cards():
     # Select * from cards;
-    stmt = db.select(Card
-    )#.where(db.or_(Card.status != "Done", Card.id > 2)).order_by(Card.title.desc())
+    stmt = db.select(
+        Card
+    )  # .where(db.or_(Card.status != "Done", Card.id > 2)).order_by(Card.title.desc())
     cards = db.session.scalars(stmt).all()
     # Excludes cards embedded in the use (for the exclude=["user.cards"])
     return CardSchema(many=True, exclude=["user.cards"]).dump(cards)
-    # dump() will convert a Python object into a JSON object, 
-    # dumps() encodes a Python object into JSON string, 
+    # dump() will convert a Python object into a JSON object,
+    # dumps() encodes a Python object into JSON string,
     # which makes it more readily readable to more frameworks/languages.
 
 
 # @cards_bp.route("/")
 # @jwt_required()
 # def all_cards():
-#     admin_required()
+#     authorize()
 #     # Select * from cards;
 #     stmt = db.select(Card)#.where(db.or_(Card.status != "Done", Card.id > 2)).order_by(Card.title.desc())
 #     cards = db.session.scalars(stmt).all()
 #     return CardSchema(many=True).dump(cards)
-#     # dump() will convert a Python object into a JSON object, 
-#     # dumps() encodes a Python object into JSON string, 
+#     # dump() will convert a Python object into a JSON object,
+#     # dumps() encodes a Python object into JSON string,
 #     # which makes it more readily readable to more frameworks/languages.
 
 
@@ -41,7 +44,7 @@ def all_cards():
 @cards_bp.route("/<int:id>")
 @jwt_required()
 def one_card(id):
-    stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
+    stmt = db.select(Card).filter_by(id=id)  # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
         # Prints user to terminal
@@ -50,22 +53,23 @@ def one_card(id):
         return CardSchema().dump(card)
     else:
         return {"error": "card not found"}, 404
-    
+
 
 # Create a new card
 @cards_bp.route("/", methods=["POST"])
 @jwt_required()
 def create_card():
     card_info = CardSchema(exclude=["id", "date_created"]).load(request.json)
-    card =  Card(
-        title = card_info["title"],
-        description = card_info.get("description", ""),
-        status = card_info.get("status", "To Do"),
-        user_id = get_jwt_identity()
+    card = Card(
+        title=card_info["title"],
+        description=card_info.get("description", ""),
+        status=card_info.get("status", "To Do"),
+        user_id=get_jwt_identity(),
     )
     db.session.add(card)
     db.session.commit()
     return CardSchema().dump(card), 201
+
 
 # Update a card
 @cards_bp.route("/<int:id>", methods=["PUT", "PATCH"])
@@ -75,6 +79,7 @@ def update_card(id):
     stmt = db.select(Card).filter_by(id=id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         card.title = card_info.get("title", card.title)
         card.description = card_info.get("description", card.description)
         card.status = card_info.get("status", card.status)
@@ -82,23 +87,18 @@ def update_card(id):
         return CardSchema().dump(card), 200
     else:
         return {"error": "card not found"}, 404
-    
 
 
 # Delete a card
 @cards_bp.route("/<int:id>", methods=["DELETE"])
 @jwt_required()
 def delete_card(id):
-    admin_required()
-    stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
+    stmt = db.select(Card).filter_by(id=id)  # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         db.session.delete(card)
         db.session.commit()
         return {}, 200
     else:
         return {"error": "card not found"}, 404
-    
-
-
-cards_bp.register_blueprint(comments_bp)
